indenting fixes.
split up msgbase into message_file and read_message

cleaned up read_message1 to make it more readable and easier to split
up for other message base types.

renamed read_message to read_post since it reads a post and not
a generic message.

make inmsg return a boolean if the message was successful
instead of setting the stored_as to 0xffffffff