add step to push to origin
When you use the CLI, after you checkout and create a new branch you need to push that to origin which let's github know about your branch. So if you want you can do the PRs in their site. This is also handy if you are using github branches to stage changes to dev and test environments otherwise your branches only exist locally.