daemonize wwivd
look for environment variable WWIV_DIR to set the bbs dir
WWIV_USER=username will set the username to change to, if not set
we'll use 'wwiv'

also properly daemonize the binary. for now we're keeping stderr
going to stderr, use systemd to control what you want done with
that. This means we detach from the terminal on startup in all
cases.