start of support for a fido style callout.net
It'll be called fido_callout.json and live in the network dir
for each network type. It'll be a JSON map of the FTN address
as the key and fido_packet_config_t as the value (you can
see what that looks like since it's embedded in networks.json
already).

If a value is nondefault in this new file, it'll eventually
override the network defaults that come in from networks.json.

Also a trivial editor will be provided in init to configure it.