experimental feature, scrolling message reader.
Add "EXPERIMENTAL_READ_PROMPT = Y" to the [WWIV] section
of wwiv.ini. It's not a final product at all, but just an
experiment of using a scrolling message reader to read a message.

You can use arrow keys up and down to scroll the message.  Pressing
enter will put you in the normal message scan prompt.  It's just
a proof-of-concept, feedback welcome.