added help to the full screen message reader.
Added an option in Defaults to toggle it on or off.  It's only
even an option if it's enabled in wwiv.ini (same tag).

Improved the UI a bit, added a help file, the help for the reader
just lists Q for quit and ? for help.  Make sure you put
mbfsed.msg into your bbs/gfiles directory since this is the help
file.