got rid of WApplication.
This is in preparation for WSession to become the new
Application.  I used to have a split but it wasn't really
right, so merged most of both of them into WSession.  Well the problem
is that WSession really is a session for executing the bbs, not
really a "user session", which I think is really what we need This way
we can cleanup after a user is finished (i.e. from after GotCaller() until
the end of the loop.

move wfc into it's own class

{G,S}etWfcStatus -> at_wfc

cleanup