add warning in FS reader when you get near timeout.
Added ability to pass a callback function to bgetch_event so you
may be notified when you are at 1 minutes left.

Changed the non-sysop timeout from 10 minutes to 5 minutes.  This
seems more than long enough (since all we have is telnet sessions
now, no need to have a short one for modem).

Moved the code to figure out timeout into 1 place so we can make
it configurable in INIT and also make it shorter during the logon
phase