Finish ^AREPLY line supoprt (FTS-0009)
When converting text from WWIV format to FTN format look for
^D0 lines and convert the known ones to FTN format.  So far only
MSGID and REPLY lines will be converted.

Also drop other ^D0 lines that we don't know about.  While at it
right now we convert all ^Dn (n>0) lines over without the ^Dn
in it (so this means wwiv taglines, if available) will make it above
the FTN tear lines, etc.

Right now the ^D0 lines move in-place, however if that's problematic
for some other software, we could do a 2 pass conversion and coalesce
them all at the top of the message if needed.