issue992 connection_type fix
Apparently, linux didn't like the lambda function being used for the HandleAccept.  Tracked down that the connection_type was being recorded correctly on the initial connection, but once it got to std::thread cleint(f), the "by reference" it was using was not the expected one.  Changed the lambda to explicitly use the value for connection_type rather than by reference.

Incidentally, this also fixed SSH connections because ALL connections were using the ConnectionType::TELNET enumueration.