hide FTN kludge lines AREA and SEEN-BY in wwiv.
Also make sure to put them back in the right format when we convert
the message text back into FTN format.

Right now they are masked as normal wwiv control lines ^D0.

Later on I'll add a setting in the message reader to show all control
lines in a message (wwiv, ftn or otherwise). But for now if you want
to see them, use wwivutil messages dump to display them in a message.

Part of bug #1055