some cleanups while looking into bugs.
While looking into 1039, I realized there was no way to delete
a user in the init based user editor.  So to fix that, the code to
delete a user needs to live in the SDK module, and not depend on
the BBS.  This is the 1st cleanup (make the BBS use existing SDK
code) towards that.