Don't allow passing SocketConnection by value.
Since we keep state about the connection in SocketConnection in
addition to just the handle, we shouldn't fork the connection
by allowing it to be passed by value.

The only place where we could have done that is in the Http code, which
is barely used, and wasn't forking it here anyway.  The change to
disallow the copy constructor and assignment operator ensures that we
won't make a copy of it.