Only use internal FTN in binkp for 5.x
If you are using networkb with 4.xx, don't present the FTN
addresses to binkp since we don't know them.  This was causing
problems on the newer builds for the weather station.

Also set more of the canonical fields in tests to make the test
environment more realistic. This was needed to test the
Config::is_5xx_or_later() method that is used to hide features from
4.x.

Also be safer when constructing a malformed FidoAddress in networkb
so that we can log an error and skip it vs. crashing the binary.