Fix the auto-purge old messages in network2.
Always use the version of Open and Create that take a
subboard_t vs. sub filename.  This way we set the max
message count per sub from the subboard_t struct, otherwise
we rely on the caller setting them explicitly (which wasn't
happening in the network2 code, hence this bug).

Also don't use the covariant return type unless we really
need it, it's cleaner to keep talking to the high level
API from a caller's perspective if possible.