make wwIV_TEST_TEMPDIR optional.
Also there's no reason to make the subdirs absolute in either
init nor the tests, especially since the wwiv legacy datastructures
have limits space for paths (81 bytes).  We already handle relative
paths in init and absolute it when the bbs is launched, so let's do
the same in the tests and also in init again.