more cleanup to irt_name so we have less globals.
Realized the logic for GetQuoteInitials was quite complex, so
created a (failing) unit test for it to see how it works.  Looks
like it does poorly with leading, trailing and embedded spaces.

Going to simplify that a bit now that I know the expected
outcomes with common wwiv strings and will make it a bit more
resilient to funky name formats.