Added --network_editor switch to wwivconfig.
This let's wwivconfig open up the network editor on an older
version of WWIV (4.x).  This will read the network.dat
and create  network.json on save.  The network.json is used by
wwiv5.x software to include the FTN information, and the
network.dat format is the same as 4.24 and later.

Also don't auto-upgrade the config.dat since we now have a
real usecase for using wwivconfig with 4.x wwiv.

The user editor should also work assuming the userrec size
hasn't changed in 4.3 (and I don't think it did), and the
4.x version never changed the offsets, etc.