Allow unknown arguments when parsing log arguments
Since we have our own logger, we started using the CommandLine class
twice, once to parse the cmdline for logger arguments and then later
for normal program arguments but this caused an error for unknown
arguments (program arguments) when parsing for just the logginng
arguments.  This lets us tell the command line parser to ignore
unknown arguments when getting the vebose (--v=xxx) argument.