Hopefully fix duplicate entries in contact.net
Also options --save and --backup to wwivutil net contact.  These
will rewrite the files (w/o duplicate entries) and also optionally
make backups of the files.

Fixed the Contact class so that it doesn't create duplicate entries
for WWIVnet style classes when asked to ensure a node exists using
the fake ftn address for it (as networkb does).

Also load the entries into a map so that they (a) stay sorted and (b)
also we don't keep propagating duplicate entries since they'll get
merged on load.