Add GUI editor for callout.net
We already have this for the fido_callout.json file, so also add
the ability to edit callout.net with a forms based editor.

Also we'll rewrite and format callout.net on save, and make a
backup copy of it.

Also added a new method backup_file(string|File) that will make
a backup of a file, so we can do this consistently throughout wwiv.