Try to get execing network commands to work right in wwivd
Since our path is set by su and not bash, we don't have
(or can't guarantee) the bbs dir is in the path, so exeute
the network commands explicitly from networkb (since we're
possibly called from wwivd).  That's what networkc already
did.

Also don't crash when writing to a closed textfile.

Display the right contact time when displaying why were
making an attempt