Add beginday support to wwivd.
This was the last thing we needed to run the bbs for (or have a
cron job scheduled to handle).  It's enabled by default and just
runs "./bbs -e". It's also configurable in wwivconfig.

Also cleaned up the wwivd page, grouped the http options near the
top to be closer to the other port numbers.  Right justified the
labels to match the other newer pages.