Fix quoting again
Looks like it was broken as part of the bgetch cleanup, so this time
the quote code is mostly implemented in the internal editor and not
inside of bgetch.  Much cleaner and much less code.

Also added some comments around some places I think we can clean
up more code since we're going to stop messing with the keyboard
buffers in such as wonky way.