Fixed double adding a single command by name.
That caused all sorts of problems, also realized how we were
using a unique_ptr and raw ptr on the object in different
plaaces made no sense when we could use a shared_ptr and let
both containers have a reference.  Much cleaner and less likely
to run into problems on exit.