Treat ignore max values of 0
If you have a conference with a max SL of 0, ignore that
constraint vs. not letting anyone in.  Ditto for age and DSL

This was causing the transfer section to not be accessible
to people unless they edited the conference that's created
by default to set the max{sl,dsl,age} to something high (like 255)