Rename ranadom_screen to printfile_random
Also moved it to a common place so it can be used elsewhere.
Cleaned up some obtuse and unneeded logic in determining if
we should call printfile or not, this was probably needed before
some simplifications in printfile years ago, but is no longer
needed since printfile will do the right thing in most cases
now.

Fixes #1182