Don't use fmtlib on linux for "{} #{}". It crashes.
not sure why, can't tell what's going on at the callstack
but it's seg faulting on linux, but not on windows.  No need
to use fmtlib here, can just StrCat it, so we'll do that.

fmtlib is updated now to v6, so will try to upgrade to see if
that fixes it, otherwise will stay with the StrCat solution.

Fixes #1196