Fix output display problem for uint8_t in UserEditor.
Some of the data fields that are uint8_t render as a character
datatype for StrCat since there's no way to tell if it's an
unsigned char that we use as a character type in WWIV or a
really short unsigned integer type.  Cast these to unsigned
so they render as numbers.

Fixes #1199