Experimental support for pipe codes in the fsed.
When editing a line, you see pipe codes expanded as |#7 or |09
however they are converted to colors and the text colorized
for all other lines other than the one you are editing.

This could work for WWIV color codes too, where we can show the
heart code for lines you are editing then every other line shows
the colors.

Maybe can do better with the WWIV color codes, but not super likely
with the pipe codes.