Add wwiv color support into the editor.
This mean't needing to track color code and also the character
for each position, so just using a string didn't work anymore.

introduced a cell_t type that has a wwiv color and character, and
track that for each character entered.  Added Control-P + # to
the FSED to set the color for the line, then when adding a new
character, use the line's current color.  This means that each
line starts off with the default color, but this is also how the
internal editor worked and seems to make sense.