Cleanup the MVC of the editor and controller.
Right now, almost all of the real key handling was happening in the
controller (fsed(...) method).  Moved most of that into the model
which responds to commands by interpreting them there.

Added another callback to the FsedView class to draw the currentline
and switched to using that.

Made the FsedView into a shared_ptr so it can be safely shared across
the model and controller.

Cleaned up some code to always gotoxy through the view and advance
the curli only from the model.