Introduce some safety with Networks.at
Also be more intentional on accessing out of bands networks
in readmail.

This is where we tend to see lots of places where wwiv accesses networks
that are out of range, in part because the network lists changed or
messages just had junk in them.  If we figure out some good patterns
for safe access here, can apply these elsewhere.