Reset the SIGCHLD signal handler on networkc and networkf.
Since wwivd ignores this signal so that orphaned BBS binaries
are not left as zombies, it may follow through to child
processes as well.  That was happening here on some linux
flavors, so we'll be explicit about resetting it to the
default early in execution of these network apps.

Fixes #1297