Added integrated support for NetFoss.
Here's how to use it:

* Download netfoss and unzip it to \wwiv\netfoss (assuming that \wwiv is
  where wwiv is installed).  if you want it in some other directory you can
  use the wwiv.ini directive NETFOSS_DIR to pick a different location.
* Edit a chain, adn change the exec mode to "Integrated NetFOSS"
  example:
  F) Exec Mode:     Integrated NetFOSS

That's it, you do not need to edit the commandline to specify a path to the
NF.BAT file, nor edit it for your environment. WWIV will create a nf.bat in
the node's temp directory every time it invokes a chain or editor using the
integrated netfoss support. WWIV will execute "nf.bat <commandline entered>"