Fixed min_k setting for FTN network auto callout
Added helper to find the bytes_waiting for all BSO files and
include that into the net_contact_rec that wwivd uses to determine
if a callout is needed for this node.

Note that this still doesn't pay attention to outbound priorities,
and includes all bundle statuses (crash, normal) the same.