Added yourinfo.msg to the list of gfiles/ wwiv uses.
This will replace the internal YourInfo hardcoded info that is displayed
if this file exists, otherwise we'll fall back to the hardcoded one.

Also needed to add @z macro for U/D ratio to make this work, and therefore
moved it into User so it can be used by the macro code.