Full screen reader not displaying color right when previous line ends in ^A (#1399)
Fix line wrap, was eating the last char at the wrap position, normally that is a space
and it works ok, but fails when that isn't the case.

When wrapping, if the we have a space carried over from the previous line, eat it.
Also when reattributing quotes, do the same to keep from having a double space after
the quote identifier and text.

Fixes #1399