Don't return a new Networks class, just rather const ref to the existing one.
This can fix issues where you do something like

auto& net = config()->networks()->neworks();

The config networks() is now a reference to the existing networks.
Therefore a reference to the vector<net_network_rec> returned from it
will be pointing to base one, not a new copy that goes out of scope
with net.