Untested support for Named Pipes vs. sockets for the BBS
Eventually WWIVd will support writing to a named pipe to talk
with the BBS nodes in addition to passing a socket handle.  This
should simplify things on OS/2.

Use -XP for (PIPE) instead of -XS or -XT to use it.  -H is obsolete
when using PIPE since the pipe name is hard coded to be
\\.\PIPE\WWIV# and \\.\PIPE\WWIV#C where # is the node number.