ScopeExit should act like GSL final_act/finally (#1375)
Add a finally(...) method to construct the ScopeExit, also let
ScopeExit take a template parameter so it's not bound to std::function<void>
but default to that if nothing else is available.

Keep our implementation so we can use swap() as that's super useful for how WWIV
needs to use this functionality.

Fixes #1375