//QWK menu for sysop doesn't list (*) config menu #1493
Don't use a menu msg file for the //QWK menus, just let wwiv autogenerate
the menu text, now that it is able to do so nicely.

Also update the default wwiv menu for QWK to look more like the text file
did, single column, updated colors, and title in a litebar.

Fixes #1493