Get rid of script_{in,out} and wwiv::endl
script_{in,out}() is no longer needed now that most my_basic
functions take the mb_interpreter_t struct as parameters and we stick
pointers to the in/out class as part of WWIV's userdata there.

wwiv::endl was only needed with WWIV's iostreams implementation, which
no longer exists, so wwiv::endl is no longer useful and is deleted.