/s not working in fsed (#1378)
Display the menu when '/' is entered in the 1st position of the line just
like we do if ESC is pressed anywhere in the FSED.

If there's folks who hate this, then it's possible to add it as an option in
(D)efaults for the users, but since / has been this way forever, I doubt there
will be any issues.

Maybe if I ever get around to configurable keymaps for fsed, it'll be an option
there.