Add max age option for importing FTN messages (#1423)
In wwivconfig, one may now configure the max number of days a packet
may be old and still imported into the bbs.  This should help with
dupes from old messages being exported, reespecially those that may
predate the dupe database.

A value of 0, which is the default, means there is no max age and
messages of any age may be imported.

Fixes ##1423