Remove need for adding empty columns to EditLabels.
Now EditLabels will do it if needed internally so that the cells
will be drawn in the right location.

Optionally re-doing the data structures to just track the last
column vs. counting columns may be better, but unless we have
aligned column widths, the rendered UI may be a bit wonky.