Fix file::set_length from failing when the file is open.
We'll try to old way (chsize/ftruncate) when the file is open and the new
way (std::filesystem::resize_file) when it's not.  That'll always work.

Also be more aggressive about creating files as needed in the sdk
message api, that's how the bbs operates so the sdk should too.

Always show the list_len in the packet dump.

Fixes #1558
Fixes #1559