Make Xfer menu display toggleable with eXpert mode option on new installs. (#1593)
* Make fsed/view.cpp use default screen settings for max values

Assume that the default screen settings are true and use width and height values for fsed/view instead of using a max or min value. Note that if the max values of 79 and 20 are not arbitrary, we can change this to use those again.

Fixes #1578

* Make Xfer menu display toggleable with the eXpert mode option.

Make Xfer menu display toggleable with the eXpert mode option in the default
packaged menus. This can still be configured with the wwivconfig menu editor,
but the default behavior out of the box should be what is expected.
Fix #1592