Add more WWIV specific instance data for WWIV bbses.
Also added an option in wwivconfig's WWIVd settings to configure which
BBS is the WWIV BBS for this WWIVd.  If yes, then wwivd will read instance.dat
and add more data into the /status handler when using ?version=2023-05
example: curl http://localhost:8080/status?version=2023-05

The 2023-05 version is the newest and also not-stable version of the /status
handler.  Feedback welcome.