Fix crash on linux mint 22 with nodelist
Fixes #1661

The issue was that with gcc13, a missing return statement doesn't get
generated as a return from the function.  Older GCC versions seem to
add in this missing RET.

There may be more of these with gcc-13, will need to look for warnings
if they exist (I didn't see one for this).